requirements.txt
Django==5.1.1
django-environ==0.11.2
django-allauth==65.0.2